LPS-131054 Move input stream class to petra-io
