+++ modules/apps/document-library/document-library-video/src/main/java/com/liferay/document/library/video/internal/converter/DLVideoFFMPEGVideoConverter.java
@@ -16,6 +16,7 @@ package com.liferay.document.library.video.internal.converter;
 
 import com.liferay.document.library.kernel.util.VideoConverter;
 import com.liferay.document.library.video.internal.configuration.DLVideoFFMPEGVideoConverterConfiguration;
+import com.liferay.petra.io.DeleteOnCloseFileInputStream;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
@@ -36,7 +37,6 @@ import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -240,27 +240,5 @@ public class DLVideoFFMPEGVideoConverter implements VideoConverter {
 
 	private volatile DLVideoFFMPEGVideoConverterConfiguration
 		_dlVideoFFMPEGVideoConverterConfiguration;
-
-	private static final class DeleteOnCloseFileInputStream
-		extends FileInputStream {
-
-		public DeleteOnCloseFileInputStream(File file)
-			throws FileNotFoundException {
-
-			super(file);
-
-			_file = file;
-		}
-
-		@Override
-		public void close() throws IOException {
-			super.close();
-
-			FileUtil.delete(_file);
-		}
-
-		private final File _file;
-
-	}
 
 }
