+++ modules/core/petra/petra-io/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly project(":core:petra:petra-lang")
 	compileOnly project(":core:petra:petra-nio")
 	compileOnly project(":core:petra:petra-string")
