+++ modules/core/petra/petra-io/src/main/java/com/liferay/petra/io/DeleteOnCloseFileInputStream.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.petra.io;
+
+import com.liferay.portal.kernel.util.FileUtil;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+
+/**
+ * @author Adolfo Pérez
+ */
+public class DeleteOnCloseFileInputStream extends FileInputStream {
+
+	public DeleteOnCloseFileInputStream(File file)
+		throws FileNotFoundException {
+
+		super(file);
+
+		_file = file;
+	}
+
+	@Override
+	public void close() throws IOException {
+		super.close();
+
+		FileUtil.delete(_file);
+	}
+
+	private final File _file;
+
+}
